Create function, payroll.py
so, this is a small project I am trying to do to understand how functions works in Python.
This is a code to calculate the payroll, with the rate per hour being 10.50, and the regular hours worked being 40. however, if the worker does overtime, he must be payed 1.5$ more per each extra hour he works.
now, the problem I am having is, that I am unable to get any result. I am having problems with my return, and print statements.
any help on how I should do this properly, would be very appreciated..